GRAPHICS_FINAL_PROJECT/hidden_surface-project/requirements.txt
streamlit==1.32.0
numpy==1.26.0
Pillow==10.0.0
opencv-python==4.8.0.76
pyopengl==3.1.7
matplotlib==3.7.1